virst/germanhcn_dst/requirements.txt
gensim
tensorflow
python-aiml
spacy
nltk
pandas
matplotlib
sklearn
pyzmq
bert-serving-server
bert-serving-client
